requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
langchain
langchain-google-genai
python-dotenv==1.0.0
